Rev all HTML files except index.html, and update rev references for all HTML files
