+++ lib/cookTask.js
@@ -3,6 +3,7 @@
 var path = require('path');
 var merge = require('merge-stream');
 var objectAssign = require('object-assign');
+var joinGlob = require('./joinGlob');
 
 var cookSubTaskConfig = function(config) {
   config = objectAssign({}, config);
@@ -16,13 +17,6 @@ var cookSubTaskConfig = function(config) {
   }
 
   return config;
-};
-
-var joinGlob = function() {
-  var joinedPath = path.join.apply(path.join, arguments);
-  var invPath = joinedPath.split('!');
-
-  return (invPath.length > 1 ? '!' : '') + invPath.join('');
 };
 
 var getFullSrc = function(config, rootConfig) {
