+++ tasks/rev/index.js
@@ -11,12 +11,12 @@ module.exports = function() {
       // Add md5 hashes to assets referenced by CSS and JS files
       'rev-assets',
 
-      // Rev CSS and JS files
+      // Rev CSS, JS, and HTML files
       // (this is done after assets, so that if a referenced asset hash changes,
       // the parent hash will change as well)
-      'rev-css-js',
+      'rev-css-js-html',
 
-      // Update asset references (images, fonts, etc) with reved filenames in compiled css + js
+      // Update asset references (images, fonts, etc) with reved filenames in compiled css + js + html
       'rev-update-references',
 
       // Finally, report filesizes
